숙소 인원수 검색 기능 추가

+++ TeamprojectTeam5/src/main/java/com/bitc/team5/dto/LodgeRoomDto.java
@@ -9,7 +9,7 @@ public class LodgeRoomDto {
 	private int seq;
 	private String lodgeName;
 	private String lodgeImg;
-	private int lodgeScore;
+	private double lodgeScore;
 	private String roomImg;
 	private String roomName;
 	private String roomType;
